+++ app/params/config.go
@@ -33,8 +33,7 @@ const (
 	// BaseDenomUnit defines the precision of the base denomination.
 	BaseDenomUnit = 18
 
-	// Testnet chain id
-	TestnetChainID = 1336
+	// Default chain ID definition
 	DefaultChainID = 1010
 )