+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Add missing address validation in `GetTokenfactoryDenomsByCreator` query to prevent potential crashes with malformed addresses
 - Override EVM chain ID if default
+- Change evm chain coin info mapping to always default
 
 ## v4.0.0 — 2025-08-06